refactor: jwtconfig도 auth로 이동
