+++ src/main/java/roomescape/auth/LoginMemberArgumentResolver.java
@@ -1,7 +1,7 @@
 package roomescape.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
-import jwt.JwtUtils;
+import auth.JwtUtils;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
